+++ media/lua/client/CHC_config_apply_funcs.lua
@@ -12,7 +12,8 @@ CHC_main.config_apply_funcs.process = function(values)
         list_font_size = p.onChangeListFontSize,
         show_recipe_module = p.onChangeShowRecipeModule,
         show_icons = p.onChangeShowIcons,
-        delayed_search = p.onChangeDelayedSearch
+        delayed_search = p.onChangeDelayedSearch,
+        show_traits = p.onChangeShowTraits,
     }
     if map[values.id] then
         map[values.id](inst)
@@ -23,14 +24,14 @@ end
 
 CHC_main.config_apply_funcs.onChangeListFontSize = function(inst)
     inst.updateQueue:push({
-        targetView = 'all',
+        targetViews = { 'all' },
         actions = { 'needUpdateFont', 'needUpdateObjects' }
     })
 end
 
 CHC_main.config_apply_funcs.onChangeShowRecipeModule = function(inst)
     inst.updateQueue:push({
-        targetView = 'all',
+        targetViews = { 'all' },
         actions = { 'needUpdateModRender' },
         exclude = {
             search_items = true,
@@ -41,14 +42,21 @@ end
 
 CHC_main.config_apply_funcs.onChangeShowIcons = function(inst)
     inst.updateQueue:push({
-        targetView = 'all',
+        targetViews = { 'all' },
         actions = { 'needUpdateShowIcons' }
     })
 end
 
 CHC_main.config_apply_funcs.onChangeDelayedSearch = function(inst)
     inst.updateQueue:push({
-        targetView = 'all',
+        targetViews = { 'all' },
         actions = { 'needUpdateDelayedSearch' }
+    })
+end
+
+CHC_main.config_apply_funcs.onChangeShowTraits = function(inst)
+    inst.updateQueue:push({
+        targetViews = { 'search_items', 'fav_items' },
+        actions = { 'needUpdateObjSource' }
     })
 end